% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CalculateNicheSubspeciesDifferences.R
\name{backgroundForPCA}
\alias{backgroundForPCA}
\title{Generating background points for a pca}
\usage{
backgroundForPCA(
  localities = locs_thinned,
  r = 1,
  num = (100 * nrow(localities)),
  e = Env,
  verbose = T
)
}
\arguments{
\item{localities}{Cleaned and thinned localities}

\item{r}{The radius in m with which to get background points}

\item{num}{The number of background points}

\item{e}{The environmental raster for niche}
}
\description{
This function generates background points for making a PCA of environment.
}
\examples{

loc_good_clean = cleanByEnvironment(Env, loc)
locs_thinned = spThinBySubspecies(loc_good_clean,thin.par=10,reps=1,lat.col="latitude",long.col="longitude",spec.col="assigned")
loc_thin_bgstuff = backgroundForPCA(localities=loc_good[,c("Longitude","Latitude")],r=200000,num=(100*nrow(localities)),e=Env)
}
