% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CalculateNicheSubspeciesDifferences.R
\name{backgroundPerSpecies}
\alias{backgroundPerSpecies}
\title{Generating background points for a species}
\usage{
backgroundPerSpecies(localities = loc_thin)
}
\arguments{
\item{localities}{Thinned cleaned localities.}
}
\description{
This function generates background points for each species.
}
\examples{

loc_good_clean = cleanByEnvironment(Env, loc)
locs_thinned = spThinBySubspecies(loc_good_clean,thin.par=10,reps=1,lat.col="latitude",long.col="longitude",spec.col="assigned")
loc_thin_bgstuff = backgroundForPCA(localities=loc_good[,c("Longitude","Latitude")],r=200000,num=(100*nrow(localities)),e=Env)
perspecies_bgstuff = backgroundPerSpecies(loc_thin)
}
