% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleanByEnvironment.R
\name{cleanByEnvironment}
\alias{cleanByEnvironment}
\title{Clean points by the environment}
\usage{
cleanByEnvironment(Env, loc, latname = "latitude", lonname = "longitude")
}
\arguments{
\item{Env}{The environmental variables to calculate niche from}

\item{loc}{The locality data}
}
\description{
This function removes any localities that do not have environmental data
so that the niche can be calculated. It assumes the second and third columns
are longitude and latitude, respectively.
}
\examples{

loc_good_clean = cleanByEnvironment(Env, loc)
}
