% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/closestPolygonFunction.R
\name{closestPolygonFunction}
\alias{closestPolygonFunction}
\title{Check for Subspecies Contiguity}
\usage{
closestPolygonFunction(listOfPolygons)
}
\arguments{
\item{listOfPolygons}{List of subspecies range polygons.}
}
\description{
This function takes multiple subspecies ranges as polygons and checks whether
1) polygon features are adjacent to their own subspecies, and if not, 2)
whether polygon features are closer to their own subspecies than to another.
If a subspecies feature does not meet either criteria, it is eliminated.
}
\details{
WARNING!!! Function is not working. Do not use.
}
