% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/densityRasterRemoveIntersection.R
\name{convertRelToPresAbs}
\alias{convertRelToPresAbs}
\title{Convert relative raster to presence/absence raster}
\usage{
convertRelToPresAbs(dens_relative_bool)
}
\arguments{
\item{ras}{dens_relative_bool}
}
\description{
This function converts relative values of a raster into a presence/absence
}
