% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createAssignedSubspecies.R
\name{createAssignedSubspecies}
\alias{createAssignedSubspecies}
\title{Species Occurrences to Subspecies Occurrences}
\usage{
createAssignedSubspecies(
  spp,
  subsppList,
  pointLimit,
  dbToQuery,
  method = "raster",
  quant = 0.95,
  xmin = -180,
  xmax = 180,
  ymin = -90,
  ymax = 90,
  plotIt = F,
  bgLayer = NULL,
  outputDir,
  datafile = NULL,
  epsilon = 1e-06,
  spp_epsilon = epsilon,
  subspp_epsilon = epsilon,
  restrictNominate = F,
  cleanup_nominate = T,
  num_digits_latlong = 2,
  cells_per_bgLayer = 50,
  downloadOnly = FALSE,
  ...
)
}
\arguments{
\item{spp}{Genus and species to query, as string}

\item{subsppList}{Strings of subspecies to query}

\item{pointLimit}{Maximum point limit to return for each database -- see spocc::occ}

\item{dbToQuery}{List of databases to search through -- see spocc::occ}

\item{method}{Whether to calculate overlaps with rasters or polygons}

\item{quant}{quant for density, below which points are removed. E.g., if set to 0.95, removes 95% least dense squares.}

\item{xmin}{Minimum longitude extent to clip rasters to}

\item{xmax}{Maximum longitude extent to clip rasters to}

\item{ymin}{Minimum latitute extent to clip rasters to}

\item{ymax}{Maximum latitude extent to clip rasters to}

\item{plotIt}{Whether to generate plots}

\item{bgLayer}{A background layer for generating plots}

\item{outputDir}{What directory to output to}

\item{datafile}{if already ran and saved output from spocc:occ, put file here -- default NULL}

\item{epsilon}{Parameter for anomaly detection}

\item{spp_epsilon}{Parameter for anomly detection specifically for species}

\item{subspp_epsilon}{Parameter for anomaly detection specifically for subspecies}

\item{restrictNominate}{Whether or not to restrict the nominate and remove extra points}

\item{cleanup_nominate}{Whether or not to clean up the nominate}

\item{num_digits_latlong}{The number of digits to round the coordinates to}

\item{cells_per_bgLayer}{How many cells to include per background layer}

\item{downloadOnly}{Whether to run as download occurrences only}
}
\description{
This function is a wrapper for the package and takes a species with a list of subspecies, and then
assigns points to subspecies. Points that have problems with assignment are listed as "suspect"
and points without problems are listed as "good." Also returns polygons used to make subspecies
assignments.
}
\examples{

Env = raster::stack(list.files(path='~/wc2-5/',pattern="\\\\.bil$",full.names=T))
ext = raster::extent(c(-125,-60,10,50))
Env = raster::crop(Env, ext)
bg = Env[[1]]
phainopeplaNitens = createAssignedSubspecies(spp="Phainopepla nitens",
   subsppList=c("nitens","lepida"),
   pointLimit=500,dbToQuery=c("gbif","bison","inat","ebird","ecoengine","vertnet"),
   quant=0.95,xmin=-125,xmax=-60,ymin=10,ymax=50,plotIt=T,bgLayer=bg,
   outputDir="~/project/")
suspect_occurrences = phainopeplaNitens$loc_suspect,
good_occurrences = phainopeplaNitens$loc_good,
subspecies_polygons = phainopeplaNitens$pol
}
