% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createPcaToCompare.R
\name{createPcaToCompare}
\alias{createPcaToCompare}
\title{Generating a pca for comparison across niches.}
\usage{
createPcaToCompare(loc_thin_bgstuff, perspecies_bgstuff, species, verbose = T)
}
\arguments{
\item{loc_thin_bgstuff}{Background points for all species.}

\item{perspecies_bgstuff}{Background points for each species.}

\item{species}{Species name, for plotting}
}
\description{
This function uses ecospat to create PCA across Environments to compare
niches. Requires some custom ecospat functions which are given below.
}
\examples{

loc_good_clean = cleanByEnvironment(Env, loc)
locs_thinned = spThinBySubspecies(loc_good_clean,thin.par=10,reps=1,lat.col="latitude",long.col="longitude",spec.col="assigned")
loc_thin_bgstuff = generateBackgroundForPCA(localities=loc_good[,c("Longitude","Latitude")],r=200000,num=(100*nrow(localities)),e=Env)
perspecies_bgstuff = generateBackgroundPerSpecies(loc_thin)
pcaOutput = createPcaToCompare(loc_thin_bgstuff,perspecies_bgstuffspecies)
pca_grid_clim = pcaOutput$grid_clim
}
