% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PairsOfBirdSpeciesNicheOverlap.R
\name{densityMapToPolygons}
\alias{densityMapToPolygons}
\title{Convert Density Maps to Polygons}
\usage{
densityMapToPolygons(densityMap, verbose = F)
}
\arguments{
\item{densityMap}{Raster of densities from subspeciesDensityMap() or similar}
}
\description{
This function converts density maps (or other raster files) to a polygon
}
\examples{

listFromSubspeciesOcc = subspeciesOccQuery(spp="Cardinalis sinuatus",
   subsppList=c("sinuatus","peninsulae","fulvescens"),pointLimit=100,
   dbToQuery="gbif")
labeledLoc = labelSubspecies(subsppOccList=listFromSubspeciesOcc)
locs_sin = labeledLoc[labeledLoc$subspecies=="sinuatus",]
dens_sin = subspeciesDensityMap(localities=locs_sin,quant=0.95,
   xmin=-125,xmax=-60,ymin=10,ymax=50)
densPol_sin = densityMapToPolygons(densityMap=dens_sin)
}
