% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/densityRasterRemoveIntersection.R
\name{densityRasterRemoveIntersection}
\alias{densityRasterRemoveIntersection}
\title{Remove Raster Intersections}
\usage{
densityRasterRemoveIntersection(densA, densB, verbose = F)
}
\arguments{
\item{densA}{The first raster}

\item{densB}{The second raster}

\item{verbose}{Whether to print extra statements}
}
\description{
This function finds intersections between raster and then removes them
}
\examples{

listFromSubspeciesOcc = subspeciesOccQuery(spp="Cardinalis sinuatus",
   subsppList=c("sinuatus","peninsulae","fulvescens"),pointLimit=100,
   dbToQuery="gbif")
labeledLoc = labelSubspecies(subsppOccList=listFromSubspeciesOcc)
locs_sin = labeledLoc[labeledLoc$subspecies=="sinuatus",]
dens_sin = subspeciesDensityMap(localities=locs_sin,quant=0.95,
   xmin=-125,xmax=-60,ymin=10,ymax=50)
XXXXXXXX

}
