% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CalculateNicheSubspeciesDifferences.R
\name{ecospat.grid.clim.dyn_custom}
\alias{ecospat.grid.clim.dyn_custom}
\title{Custom grid.clim.dyn}
\usage{
ecospat.grid.clim.dyn_custom(
  glob,
  glob1,
  sp,
  R,
  th.sp = 0,
  th.env = 0,
  geomask = NULL,
  removeNA = T
)
}
\description{
Custom ecospat function that accounts for NA values. See ecospat
}
