% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecospat_custom.R
\name{ecospat.niche.dyn.index_custom}
\alias{ecospat.niche.dyn.index_custom}
\title{Custom version of ecospat's niche dynamics testing}
\usage{
ecospat.niche.dyn.index_custom(z1, z2, intersection = 0)
}
\description{
Custom version of ecospat's niche dynamics testing
}
