% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecospat_custom.R
\name{ecospat.niche.equivalency.test_custom}
\alias{ecospat.niche.equivalency.test_custom}
\title{Custom version of ecospat's niche equivalency testing}
\usage{
ecospat.niche.equivalency.test_custom(
  z1,
  z2,
  rep,
  intersection = 0,
  overlap.alternative = "higher",
  expansion.alternative = "lower",
  stability.alternative = "higher",
  unfilling.alternative = "lower",
  ncores = 1
)
}
\description{
Custom version of ecospat's niche equivalency testing
}
