% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecospat_custom.R
\name{ecospat.niche.similarity.test_custom}
\alias{ecospat.niche.similarity.test_custom}
\title{Custom version of ecospat's niche similarity that can handle NA values}
\usage{
ecospat.niche.similarity.test_custom(
  z1,
  z2,
  rep,
  intersection = 0,
  rand.type = 1,
  ncores = 1,
  overlap.alternative = "higher",
  expansion.alternative = "lower",
  stability.alternative = "higher",
  unfilling.alternative = "lower"
)
}
\description{
Custom version of ecospat's niche similarity that can handle NA values
}
