% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PairsOfBirdSpeciesNicheOverlap.R
\name{labelSubspecies}
\alias{labelSubspecies}
\title{Convert Occurence List to Labeled Dataframe}
\usage{
labelSubspecies(subsppOccList, spp, subsppList)
}
\arguments{
\item{subsppOccList}{A list of subspecies occurences from subspeciesOccQuery() or similar}
}
\description{
This function converts a labeled list of subspecies occurences
into a dataframe of occurences, with a column for subspecies, using occ2df_subspeciesLabels().
The latter function works only on a single subspecies whereas this function
performs the task for all subspecies in the list.
}
\examples{

listFromSubspeciesOcc = subspeciesOccQuery(spp="Cardinalis sinuatus",
   subsppList=c("sinuatus","peninsulae","fulvescens"),pointLimit=100,
   dbToQuery="gbif")
labeledLoc = labelSubspecies(subsppOccList=listFromSubspeciesOcc)
subsppNames = unique(labeledLoc$subspecies)
}
