% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/localitiesToNicheMath.R
\name{localitiesToNicheMath}
\alias{localitiesToNicheMath}
\title{Wrapper for calculating niche math}
\usage{
localitiesToNicheMath(
  Env,
  loc,
  species,
  rep1 = 1000,
  rep2 = 1000,
  RMvalues = seq(0.5, 4, 0.5),
  fc = c("L", "LQ", "H"),
  numCores = 1,
  method = "block",
  verbose = T,
  runNicheModels = T,
  overwrite = F
)
}
\arguments{
\item{Env}{The environmental variables to compare.}

\item{species}{The species whose name to print.}

\item{locs}{The localities to be processed.}
}
\description{
This function serves as a wrapper for the rest of the functions.
}
\examples{

localitiesToNicheMath(Env,loc,species)
}
