% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PairsOfBirdSpeciesNicheOverlap.R
\name{locatePolygonPoints}
\alias{locatePolygonPoints}
\title{Point to Polygon Locator}
\usage{
locatePolygonPoints(
  test_points,
  polygonA,
  polygonB,
  crs = "+proj=longlat +ellps=WGS84",
  setcoord = TRUE,
  nameA = "A",
  nameB = "B"
)
}
\arguments{
\item{test_points}{The points to check where they fall}

\item{polygonA}{The first polygon}

\item{polygonB}{The second polygon}

\item{crs}{A coordinate reference system to assign to the polygons and the test points if setcoord = TRUE}

\item{setcoord}{Whether to set a coordinate reference system to the points and polygons}

\item{nameA}{The (subspecies) name associated with the first polygon}

\item{nameB}{The (subspecies) name associated with the second polygon}
}
\description{
This function takes a two polygons and a list of points
and determines whether the points fall within the polygons, then
adds a column for each polygon's name with a boolean of whether
the point overlaps.
}
\examples{

listFromSubspeciesOcc = subspeciesOccQuery(spp="Cardinalis sinuatus",
   subsppList=c("sinuatus","peninsulae","fulvescens"),pointLimit=100,dbToQuery="gbif")
labeledLoc = labelSubspecies(subsppOccList=listFromSubspeciesOcc)
locs = labeledLoc[labeledLoc$subspecies=="sinuatus",]
locs_sin = labeledLoc[labeledLoc$subspecies=="sinuatus",]
locs_ful = labeledLoc[labeledLoc$subspecies=="fulvescens",]
dens_sin = subspeciesDensityMap(localities=locs_sin,quant=0.95,
   xmin=-125,xmax=-60,ymin=10,ymax=50)
dens_ful = subspeciesDensityMap(localities=locs_ful,quant=0.95,
   xmin=-125,xmax=-60,ymin=10,ymax=50)
densPol_sin = densityMapToPolygons(densityMap=dens_sin)
densPol_ful = densityMapToPolygons(densityMap=dens_ful)
polyLocations = labeledLoc
polyLocations = locatePolygonPoints(test_points=polyLocations,polygonA=densPol_sin,
   polygonB=densPol_ful,crs="+proj=longlat +ellps=WGS84",setcoord = TRUE,nameA="sinuatus",nameB="fulvescens")
}
