% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/locateRasterPoints.R
\name{locateRasterPoints}
\alias{locateRasterPoints}
\title{Point to Raster Locator}
\usage{
locateRasterPoints(test_points, rasterA, nameA = "A")
}
\arguments{
\item{test_points}{The points to check where they fall}

\item{rasterA}{The Raster to look for overlap}

\item{nameA}{The (subspecies) name associated with the raster}
}
\description{
This function takes one raster and a list of points
and determines whether the points fall within the raster, then
adds a column for each raster's name with a boolean of whether
the point overlaps.
}
\examples{

listFromSubspeciesOcc = subspeciesOccQuery(spp="Cardinalis sinuatus",
   subsppList=c("sinuatus","peninsulae","fulvescens"),pointLimit=100,dbToQuery="gbif")
labeledLoc = labelSubspecies(subsppOccList=listFromSubspeciesOcc)
locs = labeledLoc[labeledLoc$subspecies=="sinuatus",]
locs_sin = labeledLoc[labeledLoc$subspecies=="sinuatus",]
locs_ful = labeledLoc[labeledLoc$subspecies=="fulvescens",]
dens_sin = subspeciesDensityMap(localities=locs_sin,quant=0.95,
   xmin=-125,xmax=-60,ymin=10,ymax=50)
dens_ful = subspeciesDensityMap(localities=locs_ful,quant=0.95,
   xmin=-125,xmax=-60,ymin=10,ymax=50)
densPol_sin = densityMapToPolygons(densityMap=dens_sin)
densPol_ful = densityMapToPolygons(densityMap=dens_ful)
polyLocations = labeledLoc
polyLocations = locateRasterPoints(test_points=polyLocations,polygonA=densPol_sin,
   nameA="sinuatus")
}
