% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mgdad.R
\name{mgdad}
\alias{mgdad}
\title{Outlier Detection}
\usage{
mgdad(localities, epsilon = 1e-04)
}
\arguments{
\item{localities}{A list of localities to check for anomalies}

\item{epsilon}{An value with which to flag anomalies with probability less than epsilon}
}
\description{
Calculates Multivariate Gaussian Distribution values for p_x
}
