% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mgdad_px.R
\name{mgdad_px}
\alias{mgdad_px}
\title{Outlier Probability}
\usage{
mgdad_px(localities)
}
\arguments{
\item{localities}{A list of localities to check for anomalies}
}
\description{
Calculates Multivariate Gaussian Distribution values for p_x
}
