% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nearestNeighborCleanup.R
\name{nearestNeighborCleanup}
\alias{nearestNeighborCleanup}
\title{Clean points via nearest neighbors}
\usage{
nearestNeighborCleanup(good_points, bad_points, k = 21)
}
\arguments{
\item{good_points}{XX}

\item{bad_points}{XX}

\item{k}{XX}
}
\description{
This function calculates unlabeled points based on their assigned
labels, assuming the assigned labels are correct.It calculates them
based on their location in latitude and longitude space.
}
\examples{

## i am an example

}
