% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PairsOfBirdSpeciesNicheOverlap.R
\name{occ2df_subspeciesLabels}
\alias{occ2df_subspeciesLabels}
\title{Convert Occurence Records to Labeled Dataframe}
\usage{
occ2df_subspeciesLabels(subsppOccList_object, subsppOccList_name)
}
\arguments{
\item{subsppOccList_object}{A list of subspecies occurences from subspeciesOccQuery() or similar}

\item{subsppOccList_name}{Names of the subspecies represented in subsppOccList_object}
}
\description{
This function converts a labeled list of subspecies occurences
into a dataframe of occurences, with a column for subspecies
It is called within labelSubspecies()
}
\examples{

listFromSubspeciesOcc = subspeciesOccQuery(spp="Cardinalis sinuatus",
   subsppList=c("sinuatus","peninsulae","fulvescens"),pointLimit=100,
   dbToQuery="gbif")
sppLocLab = occ2df_subspeciesLabels(subsppOccList_object=listFromSubspeciesOcc[[1]],
   subsppOccList_name=names(listFromSubspeciesOcc)[1])
}
