% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CalculateNicheSubspeciesDifferences.R
\name{overlap.sim.gen}
\alias{overlap.sim.gen}
\title{Custom overlap.sim.gen}
\usage{
overlap.sim.gen(repi, z1, z2, rand.type = rand.type)
}
\description{
Custom ENMtools similarity generator
}
