% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecospat_custom.R
\name{overlap.sim.gen_custom}
\alias{overlap.sim.gen_custom}
\title{Custom version of ecospat's overlap similarity that can handle NA values}
\usage{
overlap.sim.gen_custom(repi, z1, z2, rand.type = rand.type, intersection = 0)
}
\description{
Custom version of ecospat's overlap similarity that can handle NA values
}
