% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pairwiseNicheEquivalence.R
\name{pairwiseNicheEquivalence}
\alias{pairwiseNicheEquivalence}
\title{Test for pairwise niche equivalence}
\usage{
pairwiseNicheEquivalence(
  pca_grid_clim,
  rep1 = 1000,
  rep2 = 1000,
  species,
  verbose = T,
  overlap.alternative = "different",
  expansion.alternative = "different",
  stability.alternative = "different",
  unfilling.alternative = "different"
)
}
\arguments{
\item{pca_grid_clim}{The grid for the climate PCA}

\item{rep1}{The number of repetitions for niche equivalency}

\item{rep2}{The number of repeititons for niche similarity}
}
\description{
This function uses the background PCA grid for climate that was generated to
calculate pairwise niche equivalences for the species. This uses
randomization to pairwise equivalence so you must set the parameters for
randomizing.
}
\details{
The default alternative settings will test for niche conservatism.
}
\examples{

printPointsPdfSuspect(species,subspecies,bg,loc_suspect)
}
