% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pairwiseNicheOverlap.R
\name{pairwiseNicheOverlap}
\alias{pairwiseNicheOverlap}
\title{Make pairwise niche overlaps}
\usage{
pairwiseNicheOverlap(pca_grid_clim, verbose = T)
}
\arguments{
\item{pca_grid_clim}{The grid for the climate PCA}
}
\description{
This function uses the background PCA grid for climate that was generated to
calculate pairwise niche overlaps for the species.
}
\examples{

loc_good_clean = cleanByEnvironment(Env, loc)
locs_thinned = spThinBySubspecies(loc_good_clean,thin.par=10,reps=1,lat.col="latitude",long.col="longitude",spec.col="assigned")
loc_thin_bgstuff = backgroundForPCA(localities=loc_good[,c("Longitude","Latitude")],r=200000,num=(100*nrow(localities)),e=Env)
perspecies_bgstuff = backgroundPerSpecies(loc_thin)
pcaOutput = createPcaToCompare(loc_thin_bgstuff,perspecies_bgstuffspecies)
pca_grid_clim = pcaOutput$grid_clim
overlap_df  = pairwiseNicheOverlap(pca_grid_clim)
}
