% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/densityRasterRemoveIntersection.R
\name{presAbs}
\alias{presAbs}
\title{Calculate presence/absence raster from two density rasters}
\usage{
presAbs(densA, densB)
}
\arguments{
\item{densA}{The first raster}

\item{densB}{The second raster}
}
\description{
This function turns two density rasters into a presence/absence boolean
}
