% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/densityRasterRemoveIntersection.R
\name{recurseNeighbors}
\alias{recurseNeighbors}
\title{Recursively find neighbors}
\usage{
recurseNeighbors(cell, visited, ras, allCells, verbose = F)
}
\arguments{
\item{cell}{The cell you are checking}

\item{visited}{Cells you have already visited}

\item{ras}{The raster you are checking}

\item{allCells}{All possible valid cells}

\item{verbose}{Whether or not to print extra statements}
}
\description{
This function finds neighbors of a cell recursively
}
