% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/densityRasterRemoveIntersection.R
\name{relOverlap}
\alias{relOverlap}
\title{Calculate relative overlap raster from two density rasters}
\usage{
relOverlap(densA, densB)
}
\arguments{
\item{densA}{The first raster}

\item{densB}{The second raster}
}
\description{
This function calculates the relative overlap from two density rasters
}
