% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spThinBySubspecies.R
\name{spThinBySubspecies}
\alias{spThinBySubspecies}
\title{Thinning points by subspecies}
\usage{
spThinBySubspecies(
  loc_good_clean,
  thin.par = 10,
  reps = 1,
  lat.col = "latitude",
  long.col = "longitude",
  spec.col = "assigned",
  verbose = T,
  write.files = T,
  overwrite = F,
  species
)
}
\arguments{
\item{loc_good_clean}{Cleaned localities with associated Env data}

\item{thin.par}{The amount of km points should not be closer than}

\item{reps}{How many times to repeat the thinning}

\item{lat.col}{the column name for latitude}

\item{long.col}{The column name for longitude}

\item{spec.col}{The column name for species}
}
\description{
This function takes cleaned localities and applies a thinning algorithm so
that points are no closer than the thin.par km. Does this per species, or
whatever level of taxonomy is represented in the species column
}
\examples{

loc_good_clean = cleanByEnvironment(Env, loc)
locs_thinned = spThinBySubspecies(loc_good_clean,thin.par=10,reps=1,lat.col="latitude",long.col="longitude",spec.col="assigned")
}
