% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PairsOfBirdSpeciesNicheOverlap.R
\name{subspeciesDensityMap}
\alias{subspeciesDensityMap}
\title{Generate Subspecies Density Maps}
\usage{
subspeciesDensityMap(
  localities,
  quantile = 0.95,
  xmin = NULL,
  xmax = NULL,
  ymin = NULL,
  ymax = NULL,
  total_range
)
}
\arguments{
\item{localities}{Labeled localities as generated from labelSubspecies() or subspeciesOccQuery()}

\item{quantile}{Quantile for density, below which points are removed. E.g., if set to 0.95, removes 95% least dense squares.}

\item{xmin}{Minimum longitude extent to clip to}

\item{xmax}{Maximum longitude extent to clip to}

\item{ymin}{Minimum latitute extent to clip to}

\item{ymax}{Maximum latitude extent to clip to}
}
\description{
This function uses 2D kernel density estimation to make a density raster
of points for each subspecies. It then removes all but the most
dense cells as definied by quantile.
}
\examples{

listFromSubspeciesOcc = subspeciesOccQuery(spp="Cardinalis sinuatus",
   subsppList=c("sinuatus","peninsulae","fulvescens"),pointLimit=100,
   dbToQuery="gbif")
labeledLoc = labelSubspecies(subsppOccList=listFromSubspeciesOcc)
locs = labeledLoc[labeledLoc$subspecies=="sinuatus",]
dens = subspeciesDensityMap(localities=locs,quantile=0.95,
   xmin=-125,xmax=-60,ymin=10,ymax=50)
}
