% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subspeciesOccQuery.R
\name{subspeciesOccQuery}
\alias{subspeciesOccQuery}
\title{Pull Subspecies Occurrences}
\usage{
subspeciesOccQuery(
  spp,
  subsppList = NULL,
  pointLimit = 500,
  dbToQuery = c("gbif", "bison", "inat", "ecoengine", "vertnet"),
  ...
)
}
\arguments{
\item{spp}{Genus and species to query, as string}

\item{subsppList}{Strings of subspecies to query}

\item{pointLimit}{Maximum point limit to return for each database -- see spocc::occ}

\item{dbToQuery}{List of databases to search through -- see spocc::occ}
}
\description{
This function uses spocc::occ to query GBIF and other
repositories for one species and N number of
subspecies, specified by the user. Returns a list of
occurrence record objects.
Note: currently only tested for N=2 and N=3 subspecies.
}
\examples{

listFromSubspeciesOcc = subspeciesOccQuery(spp="Cardinalis sinuatus",
   subsppList=c("sinuatus","peninsulae","fulvescens"),pointLimit=100,
   dbToQuery="gbif")
}
