% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PairsOfBirdSpeciesNicheOverlap.R
\name{trimPolygonsByOverlap}
\alias{trimPolygonsByOverlap}
\title{Remove Overlapping Polygons}
\usage{
trimPolygonsByOverlap(polygon, idList = NULL, intList = NULL)
}
\arguments{
\item{polygon}{Polygon to remove flagged features from.}

\item{idList}{List of flagged features to completely remove.}

\item{intList}{List of intersections to remove.}
}
\description{
This function takes polygon features flagged for removal from flagPolygonOverlap()
and removes them from the polygon. Called within polygonTrimmer().
}
\examples{

listFromSubspeciesOcc = subspeciesOccQuery(spp="Cardinalis sinuatus",
   subsppList=c("sinuatus","peninsulae","fulvescens"),pointLimit=100,
   dbToQuery="gbif")
labeledLoc = labelSubspecies(subsppOccList=listFromSubspeciesOcc)
locs = labeledLoc[labeledLoc$subspecies=="sinuatus",]
locs_sin = labeledLoc[labeledLoc$subspecies=="sinuatus",]
locs_ful = labeledLoc[labeledLoc$subspecies=="fulvescens",]
dens_sin = subspeciesDensityMap(localities=locs_sin,quant=0.95,
   xmin=-125,xmax=-60,ymin=10,ymax=50)
dens_ful = subspeciesDensityMap(localities=locs_ful,quant=0.95,
   xmin=-125,xmax=-60,ymin=10,ymax=50)
densPol_sin = densityMapToPolygons(densityMap=dens_sin)
densPol_ful = densityMapToPolygons(densityMap=dens_ful)
polygonsToRemove = (flagPolygonOverlap(subsppPoly1=densPol_sin,
   subsppPoly2=densPol_ful))
sin_polygonsToRemove = polygonsToRemove$subsppApoly_toremove
ful_polygonsToRemove = polygonsToRemove$subsppBpoly_toremove
overlapToRemove_sin = polygonsToRemove$subsppA_intToRemove
overlapToRemove_ful = polygonsToRemove$subsppB_intToRemove
sin_densityPolygon_trim = trimPolygonsByOverlap(polygon=densPol_sin,
   idList = sin_polygonsToRemove,intList=overlapToRemove_sin)
ful_densityPolygon_trim = trimPolygonsByOverlap(polygon=densPol_ful,
   idList = ful_polygonsToRemove,intList=overlapToRemove_ful)
}
