% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PairsOfBirdSpeciesNicheOverlap_3ormoreSpp.R
\name{databaseToAssignedSubspecies}
\alias{databaseToAssignedSubspecies}
\title{Species Occurrences to Subspecies Occurrences}
\usage{
databaseToAssignedSubspecies(spp, subsppList, pointLimit, dbToQuery,
  quantile = 0.95, xmin = -125, xmax = -60, ymin = 10, ymax = 50,
  plotIt = F, bgLayer, outputDir, ...)
}
\arguments{
\item{spp}{Genus and species to query, as string}

\item{subsppList}{Strings of subspecies to query}

\item{pointLimit}{Maximum point limit to return for each database -- see spocc::occ}

\item{dbToQuery}{List of databases to search through -- see spocc::occ}

\item{quantile}{Quantile for density, below which points are removed. E.g., if set to 0.95, removes 95% least dense squares.}

\item{xmin}{Minimum longitude extent to clip rasters to}

\item{xmax}{Maximum longitude extent to clip rasters to}

\item{ymin}{Minimum latitute extent to clip rasters to}

\item{ymax}{Maximum latitude extent to clip rasters to}

\item{plotIt}{Whether to generate plots}

\item{bgLayer}{A background layer for generating plots}

\item{outputDir}{What directory to output to}
}
\description{
This function is a wrapper for the package and takes a species with a list of subspecies, and then
assigns points to subspecies. Points that have problems with assignment are listed as "suspect" 
and points without problems are listed as "good." Also returns polygons used to make subspecies
assignments.
}
\examples{

Env = raster::stack(list.files(path='~/wc2-5/',pattern="\\\\.bil$",full.names=T))
ext = raster::extent(c(-125,-60,10,50))
Env = raster::crop(Env, ext)
bg = Env[[1]]
phainopeplaNitens = databaseToAssignedSubspecies(spp="Phainopepla nitens",
        subsppList=c("nitens","lepida"),
        pointLimit=500,dbToQuery=c("gbif","bison","inat","ebird","ecoengine","vertnet"),
        quantile=0.95,xmin=-125,xmax=-60,ymin=10,ymax=50,plotIt=T,bgLayer=bg,
        outputDir="~/project/")
suspect_occurrences = phainopeplaNitens$loc_suspect,
good_occurrences = phainopeplaNitens$loc_good,
subspecies_polygons = phainopeplaNitens$pol
}
