% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detectSpatialOutliers.R
\name{detectSpatialOutliers}
\alias{detectSpatialOutliers}
\title{Outlier Detection}
\usage{
detectSpatialOutliers(localities, fulldataset = localities, epsilon = 1e-04)
}
\arguments{
\item{localities}{A list of localities to check for anomalies}

\item{fulldataset}{The full dataset this is taken from}

\item{epsilon}{An value with which to flag anomalies with probability less than epsilon}
}
\description{
This does outlier detection on points using Multivariate Gaussian
Distribution anomaly detection. Based off of
}
