% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecospat_custom.R
\name{ecospat.grid.clim.dyn_custom}
\alias{ecospat.grid.clim.dyn_custom}
\title{Custom version of ecospat's function that allows you to remove NA values.}
\usage{
ecospat.grid.clim.dyn_custom(
  glob,
  glob1,
  sp,
  R,
  th.sp = 0,
  th.env = 0,
  geomask = NULL,
  removeNA = T
)
}
\description{
Custom version of ecospat's function that allows you to remove NA values.
}
