% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CalculateNicheSubspeciesDifferences.R
\name{ecospat.niche.equivalency.test_custom}
\alias{ecospat.niche.equivalency.test_custom}
\title{Custom niche.equivalency.test}
\usage{
ecospat.niche.equivalency.test_custom(
  z1,
  z2,
  rep,
  alternative = "higher",
  ncores = 1
)
}
\description{
Custom ecospat equivalency that accounts for NA values. See ecospat
}
