% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecospat_custom.R
\name{overlap.eq.gen_custom}
\alias{overlap.eq.gen_custom}
\title{Custom version of ecospat's overlap equivalency that can handle NA values}
\usage{
overlap.eq.gen_custom(repi, z1, z2, intersection = 0)
}
\description{
Custom version of ecospat's overlap equivalency that can handle NA values
}
